add a script to import the UMAI previous runs
